Update README.md
This Password Generator project allows users to generate strong and secure passwords based on custom criteria. Users can choose the desired password length and whether to include uppercase letters, lowercase letters, numbers, and special characters. This helps in creating complex passwords that are hard to crack, increasing the overall security of accounts and sensitive information. Built with HTML, CSS, and JavaScript, it's fully responsive and beginner-friendly.